use-pattern: Wrap `CodegenOperation` to add extra `pattern` field
There is definitely some logic that could be moved into this new class,
but I left the change minimal because I'm not sure yet if this will be
moved up to a more general location (i.e. outside the angular
generator) and didn't want to add angular/typescript specific changes.